support floats

+++ src/eval.rs
@@ -1,7 +1,7 @@
 use crate::lexer::Token;
 use std::collections::HashMap;
 
-pub fn eval(tokens: Vec<Token>) -> u32 {
+pub fn eval(tokens: Vec<Token>) -> f32 {
     let mut precedence: HashMap<char, u8> = HashMap::new();
     precedence.insert('+', 100);
     precedence.insert('-', 100);
@@ -10,7 +10,7 @@ pub fn eval(tokens: Vec<Token>) -> u32 {
     precedence.insert('^', 230);
 
     let mut operators: Vec<char> = Vec::new();
-    let mut operands: Vec<u32> = Vec::new();
+    let mut operands: Vec<f32> = Vec::new();
 
     for token in tokens {
         match token {
@@ -31,7 +31,7 @@ pub fn eval(tokens: Vec<Token>) -> u32 {
                         '-' => operands.push(left - right),
                         '*' => operands.push(left * right),
                         '/' => operands.push(left / right),
-                        '^' => operands.push(left.pow(right)),
+                        '^' => operands.push(left.powf(right)),
                         _ => (),
                     };
 
@@ -52,7 +52,7 @@ pub fn eval(tokens: Vec<Token>) -> u32 {
             '-' => operands.push(left - right),
             '*' => operands.push(left * right),
             '/' => operands.push(left / right),
-            '^' => operands.push(left.pow(right)),
+            '^' => operands.push(left.powf(right)),
             _ => (),
         };
     }
